refactor the --resident run option into a separate file
