+++ packages/flutter_tools/lib/src/commands/build_aot.dart
@@ -13,8 +13,8 @@ import '../base/utils.dart';
 import '../build_info.dart';
 import '../dart/sdk.dart';
 import '../globals.dart';
+import '../run.dart';
 import '../runner/flutter_command.dart';
-import 'run.dart';
 
 const String _kDefaultAotOutputDir = 'build/aot';
